Update cast in msg.set
A while ago, MySensors changed to use int32_t
instead of long int (and similar for other data
types). This change was not applied to all examples,
causing some examples to fail compilation for
ESP8266 with the following message:

call of overloaded 'set(long int)' is ambiguous

This update changes the casts to match the new style.
This fixes
https://github.com/mysensors/MySensors/issues/1129
Thanks to https://github.com/henfri for reporting.

+++ examples/CO2Sensor/CO2Sensor.ino
@@ -95,7 +95,7 @@ void loop()
 	long co2ppm = 2 * ((duration/1000) - 2);
 	//Serial.print(co2ppm);
 	if ((co2ppm != lastAIQ)&&(abs(co2ppm-lastAIQ)>=10)) {
-		send(msg.set((long)ceil(co2ppm)));
+		send(msg.set((int32_t)ceil(co2ppm)));
 		lastAIQ = ceil(co2ppm);
 	}
